requirements.txt
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.9
defusedxml==0.4.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.9.1
django-allauth==0.24.1
django-appconf==1.0.1
django-autofixture==0.12.1
django-bootstrap3==6.2.2
django-compressor==2.0
django-forms-bootstrap==3.0.1
django-guardian==1.4.1
django-userena==2.0.1
djangorestframework==3.3.2
drf-nested-routers==0.11.1
easy-thumbnails==2.3
gunicorn==19.4.5
html2text==2014.12.29
ipython==4.1.2
ipython-genutils==0.1.0
itsdangerous==0.24
MarkupSafe==0.23
oauthlib==1.0.3
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
Pillow==3.2.0
psycopg2==2.6.1
ptyprocess==0.5.1
python-http-client==1.2.3
python3-openid==3.0.9
rcssmin==1.0.6
requests==2.9.1
requests-oauthlib==0.6.0
rjsmin==1.0.12
sendgrid==2.2.1
simplegeneric==0.8.1
six==1.10.0
smtpapi==0.3.1
static3==0.6.1
traitlets==4.2.1
whitenoise==2.0.6
